pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "log-sculptor"
version = "0.1.0"
description = "Parse unstructured logs by learning patterns automatically"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kevin McAllorum" }
]
keywords = ["logging", "parsing", "patterns", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: Text Processing",
]
dependencies = [
    "click>=8.0",
    "regex>=2023.0",
    "orjson>=3.9",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
duckdb = [
    "duckdb>=0.9",
]
parquet = [
    "pyarrow>=14.0",
]
all = [
    "duckdb>=0.9",
    "pyarrow>=14.0",
]

[project.scripts]
log-sculptor = "log_sculptor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/log_sculptor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=log_sculptor --cov-report=term-missing"

[tool.coverage.run]
source = ["src/log_sculptor"]
